rename utils to const, clean up code in Project Component
