+++ src/components/ProjectSection/index.tsx
@@ -19,6 +19,8 @@ export type Project = {
     projectImages: string[];
 }
 
+const DELAY_TIME = 100;
+
 export const ProjectSection: React.FC<ProjectSectionProps> = ({projects, sectionTitle}: ProjectSectionProps) => {
     return (
         <>
@@ -39,13 +41,17 @@ const ProjectComponent: React.FC<ProjectComponentProps> = ((current) => {
         setExpandedImage(expandedImage === index ? null : index);
     };
 
+    const handleTransition = (newExpandedImage: number) => {
+        setExpandedImage(newExpandedImage);
+        setTimeout(() => setIsTransitioning(false), DELAY_TIME);
+    }
+
     const goToPreviousImage = () => {
         if (expandedImage !== null && expandedImage > 0 && !isTransitioning) {
             setIsTransitioning(true);
             setTimeout(() => {
-                setExpandedImage(expandedImage - 1);
-                setTimeout(() => setIsTransitioning(false), 150);
-            }, 150);
+                handleTransition(expandedImage - 1)
+            }, DELAY_TIME);
         }
     };
 
@@ -53,9 +59,8 @@ const ProjectComponent: React.FC<ProjectComponentProps> = ((current) => {
         if (expandedImage !== null && expandedImage < projectImages.length - 1 && !isTransitioning) {
             setIsTransitioning(true);
             setTimeout(() => {
-                setExpandedImage(expandedImage + 1);
-                setTimeout(() => setIsTransitioning(false), 150);
-            }, 150);
+                handleTransition(expandedImage + 1)
+            }, DELAY_TIME);
         }
     };
